(refactor) Flytt oppsummeringsfelt til egen fil
